# openapi.yaml
openapi: 3.0.3
info:
  title: FitSync2 Backend API
  version: 0.1.0
  description: Initial contract generated from current backend controllers and DTOs.
servers:
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/users/me/profile:
    post:
      tags: [Profile]
      summary: Create profile for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
    get:
      tags: [Profile]
      summary: Get my profile
      responses:
        '200':
          description: User profile with latest inbody
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithProfileResponse'
  /api/users/me/inbody-records:
    post:
      tags: [Profile]
      summary: Create my inbody record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InBodyRecordMeRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InBodyRecordResponse'
  /api/users/{userId}/profile:
    get:
      tags: [Profile]
      summary: Get profile by userId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User profile with latest inbody
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithProfileResponse'
  /api/users/profiles/{profileId}/inbody-statistics:
    get:
      tags: [Profile]
      summary: Get inbody statistics
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Inbody statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InBodyStatisticsResponse'
  /api/exercises:
    post:
      tags: [Exercise]
      summary: Create exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseResponse'
    get:
      tags: [Exercise]
      summary: Get exercise list
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ExerciseCategory'
        - name: hidden
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageExerciseListResponse'
  /api/exercises/{exerciseId}:
    get:
      tags: [Exercise]
      summary: Get exercise detail
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Exercise detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseDetailResponse'
    patch:
      tags: [Exercise]
      summary: Update exercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseResponse'
    delete:
      tags: [Exercise]
      summary: Delete exercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: deleteType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeleteType'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 5
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: id,desc
...
  schemas:
...
    UserProfileRequest:
      type: object
      properties:
        gender:
          $ref: '#/components/schemas/Gender'
        birth:
          type: string
          format: date
        workoutGoals:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutGoal'
        exerciseCategories:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseCategory'
        disease:
          type: string
        height:
          type: number
          format: double
        weight:
          type: number
          format: double
        skeletalMuscleMass:
          type: number
          format: double
        bodyFatMass:
          type: number
          format: double
        bodyFatPercentage:
          type: number
          format: double
        bmi:
          type: number
          format: double
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    InBodyRecordMeRequest:
      type: object
      properties:
        weight:
          type: number
          format: double
        skeletalMuscleMass:
          type: number
          format: double
        bodyFatMass:
          type: number
          format: double
        bodyFatPercentage:
          type: number
          format: double
        bmi:
          type: number
          format: double
      required: [weight, skeletalMuscleMass, bodyFatMass, bodyFatPercentage, bmi]
    InBodyRecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    UserWithProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/UserHeaderInfoResponse'
        profileId:
          type: integer
          format: int64
        userProfile:
          $ref: '#/components/schemas/UserProfileDetailResponse'
      required: [userId, user, profileId, userProfile]
    UserHeaderInfoResponse:
      type: object
      properties:
        name:
          type: string
        hidden:
          type: boolean
      required: [name, hidden]
    UserProfileDetailResponse:
      type: object
      properties:
        gender:
          $ref: '#/components/schemas/Gender'
        birth:
          type: string
          format: date
        workoutGoals:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutGoal'
        exerciseCategories:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseCategory'
        disease:
          type: string
        height:
          type: number
          format: double
        weight:
          type: number
          format: double
        skeletalMuscleMass:
          type: number
          format: double
        bodyFatMass:
          type: number
          format: double
        bodyFatPercentage:
          type: number
          format: double
        bmi:
          type: number
          format: double
    InBodyStatisticsResponse:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/InBodySummary'
        trends:
          type: array
          items:
            $ref: '#/components/schemas/InBodyTrendElement'
      required: [summary, trends]
    InBodySummary:
      type: object
      properties:
        weightChange:
          type: number
          format: double
        weightChangeRate:
          type: number
          format: double
        muscleMassChange:
          type: number
          format: double
        muscleMassChangeRate:
          type: number
          format: double
        bodyFatMassChange:
          type: number
          format: double
        bodyFatMassChangeRate:
          type: number
          format: double
      required: [weightChange, weightChangeRate, muscleMassChange, muscleMassChangeRate, bodyFatMassChange, bodyFatMassChangeRate]
    InBodyTrendElement:
      type: object
      properties:
        date:
          type: string
          format: date-time
        weight:
          type: number
          format: double
        skeletalMuscleMass:
          type: number
          format: double
        bodyFatMass:
          type: number
          format: double
        bodyFatPercentage:
          type: number
          format: double
        bmi:
          type: number
          format: double
      required: [date]
    ExerciseRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        description:
          type: string
          maxLength: 1000
        details:
          type: object
          additionalProperties: true
        targets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseTargetRequest'
        effects:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EffectType'
        equipments:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Equipment'
        requiredMetrics:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/MetricType'
      required: [name, category, description, details, targets, effects, equipments, requiredMetrics]
    ExerciseUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        description:
          type: string
          maxLength: 1000
        details:
          type: object
          additionalProperties: true
        hidden:
          type: boolean
        newTargets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseTargetRequest'
        updateTargets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseTargetUpdateRequest'
        deleteTargets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseTargetDeleteRequest'
        effects:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EffectType'
        requiredMetrics:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/MetricType'
      required: [name, category, description, details, hidden, newTargets, updateTargets, deleteTargets, effects, requiredMetrics]
    ExerciseTargetRequest:
      type: object
      properties:
        bodyDetailPartId:
          type: integer
          format: int64
        targetRole:
          $ref: '#/components/schemas/TargetRole'
      required: [bodyDetailPartId, targetRole]
    ExerciseTargetUpdateRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        targetRole:
          $ref: '#/components/schemas/TargetRole'
      required: [id, targetRole]
    ExerciseTargetDeleteRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    ExerciseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    ExerciseListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        hidden:
          type: boolean
        detailParts:
          type: array
          items:
            $ref: '#/components/schemas/BodyDetailPartListResponse'
      required: [id, name, category, hidden, detailParts]
    BodyDetailPartListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        detailPartName:
          type: string
        partName:
          type: string
      required: [id, detailPartName, partName]
    ExerciseDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        description:
          type: string
        details:
          type: object
          additionalProperties: true
        hidden:
          type: boolean
        targets:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseTargetDetailResponse'
        effects:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EffectType'
        equipments:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Equipment'
        requiredMetric:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/MetricType'
      required: [id, name, category, description, details, hidden, targets, effects, equipments, requiredMetric]
    ExerciseTargetDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bodyDetailPart:
          $ref: '#/components/schemas/BodyDetailPartResponse'
        targetRole:
          $ref: '#/components/schemas/TargetRole'
      required: [id, bodyDetailPart, targetRole]
    BodyDetailPartResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        detailPartName:
          type: string
        partName:
          type: string
      required: [id, detailPartName, partName]
...
    PageExerciseListResponse:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ExerciseListResponse'
...
    PageBase:
      type: object
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
      required: [content, totalElements, totalPages, size, number, numberOfElements, first, last, empty]
    DeleteType:
      type: string
      enum: [SOFT, HARD]
...
    Gender:
      type: string
      enum: [MALE, FEMALE]
    WorkoutGoal:
      type: string
      enum: [WEIGHT_LOSS, MUSCLE_GAIN, BODY_BALANCE, STRENGTH, ENDURANCE, REHABILITATION, HEALTH_MAINTENANCE]
    ExerciseCategory:
      type: string
      enum: [FITNESS, CROSSFIT, YOGA, PILATES, REHAB]
    TargetRole:
      type: string
      enum: [MAIN, SUB]
    EffectType:
      type: string
      enum: [STRENGTH, ENDURANCE, CARDIO, FLEXIBILITY, MOBILITY]
    Equipment:
      type: string
      enum: [NO_EQUIPMENT, DUMBBELLS, BARBELL, BENCH, CABLE, BOSU_TRAINER, TRX, BOX, STABILITY_BALL, MEDICINE_BALL, CONES, WEIGHT_MACHINES, PULL_UP_BAR, LADDER, HURDLES, KETTLE_BELLS, ROPE]
    MetricType:
      type: string
      enum: [WEIGHT, REPS, DISTANCE, TIME, SPEED, RPE, REST_TIME]
...